new design imp

+++ Stock_Management/src/main/java/com/stock_management/dto/invoice/PurchaseInvoiceLineDto.java
@@ -3,12 +3,14 @@ package com.stock_management.dto.invoice;
 import com.stock_management.dto.stock.StockDto;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PurchaseInvoiceLineDto {
     private Long purchaseInvoiceLineId;
     private StockDto stockDto;
-    private Double wholeSalePrice;
-    private Double oldPricePerBox;
-    private Double pricePerBox;
-    private Double boxesReceived;
+    private BigDecimal wholeSalePrice;
+    private BigDecimal oldPricePerBox;
+    private BigDecimal pricePerBox;
+    private BigDecimal boxesReceived;
 }
