+++ Stock_Management/src/main/java/com/stock_management/dto/payment/PaymentDto.java
@@ -7,6 +7,7 @@ import com.stock_management.entity.Invoice;
 import com.stock_management.type.PaymentType;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,7 +15,7 @@ public class PaymentDto {
     private Long paymentId;
     private InvoiceDto invoiceDto;
     private ReceiptDto receiptDto;
-    private Double amountPaid;
+    private BigDecimal amountPaid;
     private PaymentType paymentMode;
     private UserDto createdBy;
     private Date createdDate;
