+++ Stock_Management/src/main/java/com/stock_management/dto/product/ProductQuantityPriceDto.java
@@ -2,13 +2,15 @@ package com.stock_management.dto.product;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductQuantityPriceDto {
     private String productName;
     private Integer boxesOrdered;
     private Integer unitsOrdered;
-    private Double price;
-    private Double pricePerBox;
-    private Double pricePerUnit;
+    private BigDecimal price;
+    private BigDecimal pricePerBox;
+    private BigDecimal pricePerUnit;
     private Integer unitsPerBox;
 }
