+++ Stock_Management/src/main/java/com/stock_management/dto/product/ProductStockDto.java
@@ -2,17 +2,18 @@ package com.stock_management.dto.product;
 import com.stock_management.dto.shared.AuditDto;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class ProductStockDto extends AuditDto {
     private Long stockId;
-    private Double quantity;
+    private BigDecimal quantity;
     private Integer unitsPerBox;
-    private Double wholeSalePrice;
-    private Double pricePerBox;
-    private Double pricePerUnit;
-    private Double unitsTotal;
+    private BigDecimal wholeSalePrice;
+    private BigDecimal pricePerBox;
+    private BigDecimal pricePerUnit;
+    private Integer unitsTotal;
     private LocalDate expiryDate;
-    private Double maxUnitsCanBeEntered;
+    private BigDecimal maxUnitsCanBeEntered;
 }
