+++ Stock_Management/src/main/java/com/stock_management/dto/product/SaveProductDto.java
@@ -4,6 +4,7 @@ import com.stock_management.dto.supplier.SupplierDto;
 import com.stock_management.dto.shared.AuditDto;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,11 +19,11 @@ public class SaveProductDto extends AuditDto {
     private Boolean requirePrescription;
     private String slot;
     private SupplierDto supplier;
-    private Double box;
-    private Double unitsTotal;
-    private Double wholeSalePrice;
-    private Double pricePerBox;
-    private Double pricePerUnit;
+    private BigDecimal box;
+    private BigDecimal unitsTotal;
+    private BigDecimal wholeSalePrice;
+    private BigDecimal pricePerBox;
+    private BigDecimal pricePerUnit;
     private LocalDate expiryDate;
     private Long userId;
 }
