+++ Stock_Management/src/main/java/com/stock_management/dto/receipt/ReceiptDto.java
@@ -5,14 +5,16 @@ import com.stock_management.dto.doctor.DoctorDto;
 import com.stock_management.dto.invoice.InvoiceDto;
 import com.stock_management.dto.security.UserDto;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class ReceiptDto {
     private Long receiptId;
     private InvoiceDto invoiceDto;
-    private Double totalPrice;
-    private Double discount;
+    private BigDecimal totalPrice;
+    private BigDecimal discount;
     private CustomerDto customerDto;
     private DoctorDto doctorDto;
     private UserDto createdBy;
