+++ Stock_Management/src/main/java/com/stock_management/dto/stock/UpdateStockAmountDto.java
@@ -2,9 +2,11 @@ package com.stock_management.dto.stock;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateStockAmountDto {
     private Long id;
     private Integer amount;
-    private Double newPrice;
+    private BigDecimal newPrice;
 }
